+++ tests/performance/results.js
@@ -1,27 +1,13 @@
 #!/usr/bin/env node
 
 const fs = require('fs');
+const { median } = require( './utils' );
 const testSuites = [
     'home-classic-theme',
     'home-block-theme',
 ];
 
 console.log( '\n>> 🎉 Results 🎉 \n' );
-
-/**
- * Computes the median number from an array numbers.
- *
- * @param {number[]} array
- *
- * @return {number} Median.
- */
-function median( array ) {
-	const mid = Math.floor( array.length / 2 );
-	const numbers = [ ...array ].sort( ( a, b ) => a - b );
-	return array.length % 2 !== 0
-		? numbers[ mid ]
-		: ( numbers[ mid - 1 ] + numbers[ mid ] ) / 2;
-}
 
 for ( const testSuite of testSuites ) {
     const resultsFilename = __dirname + '/specs/' + testSuite + '.test.results.json';
